delete jobs stdout/stderr files only if they exist (but I don't know why they do not already exist)

+++ python/soma_workflow/test/workflow_tests/test_fake_pipeline.py
@@ -109,8 +109,10 @@ class FakePipelineTest(WorkflowTest):
                                   "job stderr not empty : cf %s" %
                                   job_stderr_file)
                 finally:
-                  os.unlink(job_stdout_file)
-                  os.unlink(job_stderr_file)
+                  if os.path.exists(job_stdout_file):
+                    os.unlink(job_stdout_file)
+                  if os.path.exists(job_stderr_file):
+                    os.unlink(job_stderr_file)
 
 
 if __name__ == '__main__':
